+++ src/maths/Maths.h
@@ -6,12 +6,12 @@
 namespace Maths {
 	// ----- CONSTANTS -----
 
-	const float PI = 3.1415926535897932384626433832795f;
-	const float TAU = Maths::PI * 2.f;
+	const double PI = 3.1415926535897932384626433832795;
+	const double TAU = Maths::PI * 2.f;
 
-	const float ToRadians = Maths::PI / 180.f;
+	const double ToRadians = Maths::PI / 180.;
 
-	float Cubic(const float& a, const float& b, const float& c, const float& d, const float& factor);
+	double Cubic(const double& a, const double& b, const double& c, const double& d, const double& factor);
 
-	float Map(const float& v, const float& fromMin, const float& fromMax, const float& toMin, const float& toMax, const bool& clamp = false);
+	double Map(const double& v, const double& fromMin, const double& fromMax, const double& toMin, const double& toMax, const bool& clamp = false);
 };
