+++ src/maths/Vector3D.cpp
@@ -5,7 +5,7 @@
 
 #include "Vector3D.h"
 
-Vector3D::Vector3D(const float& x, const float& y, const float& z) {
+Vector3D::Vector3D(const double& x, const double& y, const double& z) {
 	m_x = x;
 	m_y = y;
 	m_z = z;
@@ -53,7 +53,7 @@ Vector3D& Vector3D::operator+=(const Vector3D& other) {
 	return *this;
 }
 
-Vector3D& Vector3D::operator/=(const float& scalar) {
+Vector3D& Vector3D::operator/=(const double& scalar) {
 	m_x /= scalar;
 	m_y /= scalar;
 	m_z /= scalar;
@@ -61,7 +61,7 @@ Vector3D& Vector3D::operator/=(const float& scalar) {
 	return *this;
 }
 
-Vector3D& Vector3D::operator*=(const float& scalar) {
+Vector3D& Vector3D::operator*=(const double& scalar) {
 	m_x *= scalar;
 	m_y *= scalar;
 	m_z *= scalar;
@@ -69,7 +69,7 @@ Vector3D& Vector3D::operator*=(const float& scalar) {
 	return *this;
 }
 
-float Vector3D::DotProduct(const Vector3D& v) const {
+double Vector3D::DotProduct(const Vector3D& v) const {
 	return (m_x * v.m_x) + (m_y * v.m_y) + (m_z * v.m_z);
 }
 
@@ -79,26 +79,26 @@ Vector3D Vector3D::CrossProduct(const Vector3D& v) const {
 		m_x * v.m_y - m_y * v.m_x);
 }
 
-float Vector3D::SqrToroidalDistance(const Vector3D& a, const Vector3D& b, const Vector3D& min, const Vector3D& max) {
+double Vector3D::SqrToroidalDistance(const Vector3D& a, const Vector3D& b, const Vector3D& min, const Vector3D& max) {
 	Vector3D delta = b - a;
 	delta.Abs();
 
 	Vector3D minMaxDelta = max - min;
 	minMaxDelta.Abs();
 
-	Vector3D mid = minMaxDelta / 2.f;
+	Vector3D mid = minMaxDelta / 2.;
 
 	if (delta.m_x > mid.m_x) delta.m_x = minMaxDelta.m_x - delta.m_x;
 	if (delta.m_y > mid.m_y) delta.m_y = minMaxDelta.m_y - delta.m_y;
 	if (delta.m_z > mid.m_z) delta.m_z = minMaxDelta.m_z - delta.m_z;
 
-	return std::fabsf(delta.SqrMagnitude());
+	return std::fabs(delta.SqrMagnitude());
 }
 
 void Vector3D::Abs() {
-	m_x = std::fabsf(m_x);
-	m_y = std::fabsf(m_y);
-	m_z = std::fabsf(m_z);
+	m_x = std::fabs(m_x);
+	m_y = std::fabs(m_y);
+	m_z = std::fabs(m_z);
 }
 
 void Vector3D::Clamp(const Vector3D& min, const Vector3D& max) {
@@ -107,24 +107,24 @@ void Vector3D::Clamp(const Vector3D& min, const Vector3D& max) {
 	m_z = std::clamp(m_z, min.m_z, max.m_z);
 }
 
-Vector3D Vector3D::Lerp(const Vector3D& min, const Vector3D& max, const float& factor, const bool& clamp) {
-	return Vector3D(Maths::Map(factor, 0.f, 1.f, min.m_x, max.m_x, clamp),
-		Maths::Map(factor, 0.f, 1.f, min.m_y, max.m_y, clamp),
-		Maths::Map(factor, 0.f, 1.f, min.m_z, max.m_z, clamp));
+Vector3D Vector3D::Lerp(const Vector3D& min, const Vector3D& max, const double& factor, const bool& clamp) {
+	return Vector3D(Maths::Map(factor, 0., 1., min.m_x, max.m_x, clamp),
+		Maths::Map(factor, 0., 1., min.m_y, max.m_y, clamp),
+		Maths::Map(factor, 0., 1., min.m_z, max.m_z, clamp));
 }
 
-float Vector3D::SqrMagnitude() const {
+double Vector3D::SqrMagnitude() const {
 	return (m_x * m_x) + (m_y * m_y) + (m_z * m_z);
 }
 
-float Vector3D::Magnitude() const {
-	return std::sqrtf(SqrMagnitude());
+double Vector3D::Magnitude() const {
+	return std::sqrt(SqrMagnitude());
 }
 
 void Vector3D::Normalize() {
-	float mag = SqrMagnitude();
-	if (mag != 1.f) {
-		mag = std::sqrtf(mag);
+	double mag = SqrMagnitude();
+	if (mag != 1.) {
+		mag = std::sqrt(mag);
 
 		m_x /= mag;
 		m_y /= mag;
@@ -134,23 +134,23 @@ void Vector3D::Normalize() {
 
 Vector3D Vector3D::Reflect(const Vector3D& vector, const Vector3D& normal) {
 	Vector3D t = Vector3D::One * normal.DotProduct(vector);
-	t *= normal * 2.f;
+	t *= normal * 2.;
 	t = vector - t;
 	t.Normalize();
 	return t;
 }
 
-Vector3D Vector3D::Refract(const Vector3D& vector, const Vector3D& normal, const float refractionRatio) {
-	Vector3D vectorInv = vector * -1.f;
+Vector3D Vector3D::Refract(const Vector3D& vector, const Vector3D& normal, const double refractionRatio) {
+	Vector3D vectorInv = vector * -1.;
 
-	float cosTheta = std::fminf(1.f, vectorInv.DotProduct(normal));
+	double cosTheta = std::fmin(1., vectorInv.DotProduct(normal));
 
 	Vector3D rOutPerp = (vector + (normal * cosTheta)) * refractionRatio;
-	Vector3D rOutPara = normal * -std::sqrtf(std::fabsf(1.f - rOutPerp.SqrMagnitude()));
+	Vector3D rOutPara = normal * -std::sqrt(std::fabs(1. - rOutPerp.SqrMagnitude()));
 
-	float sinTheta = std::sqrtf(1.f - cosTheta * cosTheta);
+	double sinTheta = std::sqrt(1. - cosTheta * cosTheta);
 
-	bool cannotRefract = refractionRatio * sinTheta > 1.f;
+	bool cannotRefract = refractionRatio * sinTheta > 1.;
 
 	if (cannotRefract) {
 		return Vector3D::Reflect(vector, normal);
@@ -162,36 +162,36 @@ Vector3D Vector3D::Refract(const Vector3D& vector, const Vector3D& normal, const
 }
 
 Vector3D Vector3D::UVSphere(const Vector3D& v) {
-	return Vector3D(0.5f + (std::atan2f(v.m_x, v.m_z) / Maths::TAU), 0.5f - (std::asinf(v.m_y) / Maths::PI), 0.f);
+	return Vector3D(0.5f + (std::atan2(v.m_x, v.m_z) / Maths::TAU), 0.5f - (std::asin(v.m_y) / Maths::PI), 0.);
 }
 
-Vector3D Vector3D::RandomVector(const float& min, const float& max) {
-	return Vector3D(Random::RandomFloat(min, max), Random::RandomFloat(min, max), Random::RandomFloat(min, max));
+Vector3D Vector3D::RandomVector(const double& min, const double& max) {
+	return Vector3D(Random::RandomDouble(min, max), Random::RandomDouble(min, max), Random::RandomDouble(min, max));
 }
 
 Vector3D Vector3D::RandomInHemisphere(const Vector3D& normal) {
 	Vector3D o = Vector3D::RandomInUnitSphere();
 
-	if (normal.DotProduct(o) < 0.f) o *= -1.f;
+	if (normal.DotProduct(o) < 0.) o *= -1.;
 	return o;
 }
 
 Vector3D Vector3D::RandomInUnitDisk() {
-	Vector3D o(Random::RandomFloat(-1.f, 1.f), Random::RandomFloat(-1.f, 1.f));
+	Vector3D o(Random::RandomDouble(-1., 1.), Random::RandomDouble(-1., 1.));
 
-	if (o.SqrMagnitude() > 1.f) o.Normalize();
+	if (o.SqrMagnitude() > 1.) o.Normalize();
 	return o;
 }
 
 Vector3D Vector3D::RandomInUnitSphere() {
-	Vector3D o = Vector3D::RandomVector(-1.f, 1.f);
+	Vector3D o = Vector3D::RandomVector(-1., 1.);
 
-	if (o.SqrMagnitude() > 1.f) o.Normalize();
+	if (o.SqrMagnitude() > 1.) o.Normalize();
 	return o;
 }
 
 Vector3D Vector3D::RandomUnitVector() {
-	Vector3D o = Vector3D::RandomVector(-1.f, 1.f);
+	Vector3D o = Vector3D::RandomVector(-1., 1.);
 	o.Normalize();
 	return o;
 }
